# util: accelerate square transpose using rayon