--- a/util/Cargo.toml
+++ b/util/Cargo.toml
@@ -6,11 +6,15 @@ license = "MIT OR Apache-2.0"
 
 [dependencies]
 serde.workspace = true
+rayon = { workspace = true, optional = true }
 
 [dev-dependencies]
 rand.workspace = true
 criterion.workspace = true
 serde_json.workspace = true
+
+[features]
+parallel = ["rayon"]
 
 [[bench]]
 name = "bit_reverse"