--- a/util/benches/bit_reverse.rs
+++ b/util/benches/bit_reverse.rs
@@ -24,7 +24,7 @@ fn bench_reverse_bits(c: &mut Criterion) {
 fn bench_reverse_slice_index_bits(c: &mut Criterion) {
     let mut group = c.benchmark_group("reverse_slice_index_bits");
     let mut rng = SmallRng::seed_from_u64(1);
-    for log_size in [1, 3, 5, 8, 16, 24] {
+    for log_size in [1, 3, 5, 8, 16, 24, 26] {
         let size = 1 << log_size;
         group.bench_with_input(BenchmarkId::from_parameter(size), &size, |b, &size| {
             let data: Vec<u64> = (0..size).map(|_| rng.random()).collect();