--- a/util/src/lib.rs
+++ b/util/src/lib.rs
@@ -98,7 +98,10 @@ const SMALL_ARR_SIZE: usize = 1 << 16;
 ///
 /// If the whole array fits in fast cache, then the trivial algorithm is cache friendly. Also, if
 /// `T` is really big, then the trivial algorithm is cache-friendly, no matter the size of the array.
-pub fn reverse_slice_index_bits<F>(vals: &mut [F]) {
+pub fn reverse_slice_index_bits<F>(vals: &mut [F])
+where
+    F: Copy + Send + Sync,
+{
     let n = vals.len();
     if n == 0 {
         return;