This approach is directly inspired by https://github.com/recmo/goldilocks/blob/main/ntt/src/permute/square.rs

I noticed that currently the transposition of square matrices never uses parallelization, and we could gain by doing this for large matrices. So basically, I added a parallelization threshold to enable this parallelization when needed.
- I added a unified approach for unit tests to avoid duplicating tests and to be able to test larger dimensions in a single loop and be able to trigger the parallelization strategy.
- When the parallel cfg is disabled, it should revert exactly to the previous logic.
- I've run the benchmarks of the crate to confirm the effect on large matrices and I also added 1 << 26 to confirm this on an even bigger matrix. Here are the results:

```shell
reverse_slice_index_bits/2
                        time:   [23.612 ns 23.870 ns 24.309 ns]
                        change: [−0.3428% +1.1072% +3.2738%] (p = 0.27 > 0.05)
                        No change in performance detected.
Found 8 outliers among 100 measurements (8.00%)
  8 (8.00%) high severe
reverse_slice_index_bits/8
                        time:   [31.910 ns 31.981 ns 32.076 ns]
                        change: [−3.5554% −1.7324% −0.3097%] (p = 0.02 < 0.05)
                        Change within noise threshold.
Found 7 outliers among 100 measurements (7.00%)
  1 (1.00%) low severe
  3 (3.00%) low mild
  2 (2.00%) high mild
  1 (1.00%) high severe
reverse_slice_index_bits/32
                        time:   [50.220 ns 50.243 ns 50.268 ns]
                        change: [−1.1057% −0.5154% −0.1680%] (p = 0.03 < 0.05)
                        Change within noise threshold.
Found 5 outliers among 100 measurements (5.00%)
  1 (1.00%) low mild
  3 (3.00%) high mild
  1 (1.00%) high severe
reverse_slice_index_bits/256
                        time:   [216.02 ns 216.68 ns 217.92 ns]
                        change: [−0.8548% −0.3059% +0.1152%] (p = 0.25 > 0.05)
                        No change in performance detected.
Found 5 outliers among 100 measurements (5.00%)
  2 (2.00%) low mild
  2 (2.00%) high mild
  1 (1.00%) high severe
reverse_slice_index_bits/65536
                        time:   [134.02 µs 136.70 µs 140.57 µs]
                        change: [+1.9106% +12.805% +21.804%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 9 outliers among 100 measurements (9.00%)
  3 (3.00%) low mild
  2 (2.00%) high mild
  4 (4.00%) high severe
reverse_slice_index_bits/16777216
                        time:   [29.594 ms 29.852 ms 30.130 ms]
                        change: [−35.150% −34.000% −32.937%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 4 outliers among 100 measurements (4.00%)
  3 (3.00%) high mild
  1 (1.00%) high severe
Benchmarking reverse_slice_index_bits/67108864: Warming up for 3.0000 s
Warning: Unable to complete 100 samples in 5.0s. You may wish to increase target time to 13.4s, or reduce sample count to 30.
reverse_slice_index_bits/67108864
                        time:   [137.48 ms 140.69 ms 144.33 ms]
                        change: [−43.812% −41.142% −38.775%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 12 outliers among 100 measurements (12.00%)
  2 (2.00%) high mild
  10 (10.00%) high severe
```

I believe that as you said here https://github.com/tcoratger/whir-p3/pull/59 for some reason, case 16 is very special as the parallel computation seems to be slower for this case and I cannot really understand why but for all the other tests of high dimensions this is the opposite and we can observe speedups up to 40% for the largest test (1 << 26).

I also wrote a write up about the method here https://hackmd.io/@tcoratger/Hyu8A-KWee in addition to the documentation that I reformulated for everyone to understand this tricky algorithm in the future.