@SyxtonPrime 2025-05-20
Yep this looks pretty good. Might be sensible to also look at the transpose crate we use and see if we can merge everything together? (I think that one also doesn't use parallelisation)