@tcoratger 2025-05-20
> Yep this looks pretty good. Might be sensible to also look at the transpose crate we use and see if we can merge everything together? (I think that one also doesn't use parallelisation)

Yes I don't know what you think but I was planning to examine this in a follow up PR.

I think that in the rectangular case, this approach can work quite well https://github.com/recmo/goldilocks/blob/main/ntt/src/permute/gw18.rs

And then there is also this part that is quite nice https://github.com/recmo/goldilocks/blob/e758871222d22ac7af0cf67a08470391e95ee317/ntt/src/permute/mod.rs#L28 where we can choose a strategy as a function of the matrix we have to transpose so that for example if we call the transpose function but by chance the matrix is ​​square, then we go directly to the square strategy, saving us a lot of time.

But to avoid long reviews, maybe this is a good idea to split into multiple PRs, this one being the first, what do you think?